+++ packages/crypto/src/utils/bignum.ts
@@ -1,8 +1,8 @@
 import BigNumber from "bignumber.js";
 
 class Bignum extends BigNumber {
-    public static readonly ZERO = new BigNumber(0)
-    public static readonly ONE = new BigNumber(1)
+    public static readonly ZERO = new BigNumber(0);
+    public static readonly ONE = new BigNumber(1);
 }
 
 Bignum.config({ DECIMAL_PLACES: 0 });